backend/requirements.txt
django-allauth
Werkzeug
pyOpenSSL
Django
graphene_django
django_extensions
psycopg2-binary
django-filter